Deal with global resources
Some resources like IAM do not belong to a region so as we tried to
delete an IAM user from all regions the deletion for the first region
succeeded but the remaining attempts, on the remaining regions, failed.

So, to deal with global resources one only need to add the code on
`GetAllResources` after the initial loops that iterate over the regions.